requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.5
aiosqlite>=0.20.0
python-dotenv>=1.0.1
